Multilevel_Inheritance
Multilevel Inheritance is an OOP concept where a class (subclass or derived class) inherits from another class, and then a new class is derived from this intermediate class. It creates a chain of inheritance, allowing properties and behaviors to be passed down through multiple levels of classes, forming a hierarchy.